+++ app/Http/Requests/UpdateSite.php
@@ -39,6 +39,7 @@ class UpdateSite extends FormRequest
             'source_repo' => 'required_unless:type,redirect|nullable|url',
             'source_branch' => 'nullable|string',
             'project_root' => 'required_unless:type,redirect|nullable|string',
+            'public_dir' => 'required_if:type,laravel|nullable|string',
             'redirect_type' => 'required_if:type,redirect|nullable|integer',
             'redirect_to' => 'required_if:type,redirect|nullable|string',
             'is_enabled' => 'boolean',