+++ tests/Unit/Http/Requests/UpdateSiteRequestTest.php
@@ -96,13 +96,16 @@ class UpdateSiteRequestTest extends TestCase
             $dataWithout['type'] = $type;
             $dataWith['type'] = $type;
 
+            $dataWithout['public_dir'] = 'laravel' === $type ? '/public' : '';
+            $dataWith['public_dir'] = $dataWithout['public_dir'];
+
             $v = $this->getValidator($dataWithout);
             $this->assertEquals($error, $v->errors()->first('source_repo'));
-            $this->assertFalse($v->passes());
+            $this->assertFalse($v->passes(), "{$type} site passed validation without source_repo");
 
             $v = $this->getValidator($dataWith);
             $this->assertArrayNotHasKey('source_repo', $v->errors()->toArray());
-            $this->assertTrue($v->passes());
+            $this->assertTrue($v->passes(), "{$type} site failed validation with source_repo");
         }
     }
 
@@ -178,6 +181,52 @@ class UpdateSiteRequestTest extends TestCase
         $this->assertFalse($this->validateField('project_root', true));
         $this->assertFalse($this->validateField('project_root', ['a', 'b']));
         $this->assertFalse($this->validateField('project_root', (object) ['a', 'b']));
+    }
+
+    /** @test */
+    public function site_public_dir_is_required_when_type_is_laravel(): void
+    {
+        $v = $this->getValidator(['type' => 'laravel']);
+
+        $this->assertStringContainsString('required', $v->errors()->first('public_dir'));
+        $this->assertFalse($v->passes());
+
+        $v = $this->getValidator([
+            'name' => 'foo',
+            'primary_domain' => 'localhost',
+            'type' => 'laravel',
+            'project_root' => '/',
+            'public_dir' => '/laravel',
+            'source_repo' => 'https://github.com/foo/bar.git',
+        ]);
+
+        $this->assertEmpty($v->errors()->get('public_dir'));
+        $this->assertTrue($v->passes());
+    }
+
+    /** @test */
+    public function site_public_dir_is_not_required_when_type_is_not_laravel(): void
+    {
+        $v = $this->getValidator([
+            'name' => 'foo',
+            'primary_domain' => 'localhost',
+            'type' => 'php',
+            'project_root' => '/',
+            'source_repo' => 'https://github.com/foo/bar.git',
+        ]);
+
+        $this->assertEmpty($v->errors()->get('public_dir'));
+        $this->assertTrue($v->passes());
+    }
+
+    /** @test */
+    public function site_public_dir_must_be_a_string(): void
+    {
+        $this->assertTrue($this->validateField('public_dir', '/'));
+        $this->assertFalse($this->validateField('public_dir', 42));
+        $this->assertFalse($this->validateField('public_dir', true));
+        $this->assertFalse($this->validateField('public_dir', ['a', 'b']));
+        $this->assertFalse($this->validateField('public_dir', (object) ['a', 'b']));
     }
 
     /** @test */